requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
streamlit==1.40.0
langchain==0.3.0
langchain-openai==0.2.0
langchain-groq
langchain-community==0.3.0
langchain-huggingface==0.0.3
chromadb==0.5.4
openai==1.57.0
pydantic==2.8.2
python-dotenv==1.0.1
requests==2.32.3
sentence-transformers
langchain-huggingface
